Changed capture_mouse to be more generic (#180)
* Changed capture_mouse to triggered_menu

`capture_mouse` event occurs when spawning menu, it is not related directly to mouse. Therefore, `triggered_menu` is more general as this event can be used for various additions other than mouse capture in lua.

* Changed capture_mouse to triggered_menu

Renamed `capture_mouse` to `triggered_menu` 
`capture_mouse` event occurs when spawning menu, it is not related directly to mouse. Therefore, `triggered_menu` is more general as this event can be used for various additions other than mouse capture in lua.