Implement forcing of a codec in transcoding profile
This adds two optional strings to the `conf.json`:
- force_video_codec
- force_audio_codec

When set, these will replace VideoCodec and AudioCodec,
respectively and as-is, in the transcoding profile POSTed
to the server. Other values in `conf.json` that tweak this
profile will be overridden by `force_video_codec`.

Benefits of this is to specify the exact codecs required
in some low-end systems, which may not have hardware
h264 decoding available, thus improving performance.

Signed-off-by: Han Sol Jin <hansol@hansol.ca>